initial commit to add a reload toolkit panel button
